requirements.txt
# Server
fastapi
uvicorn[standard]
python-dotenv

# Pydantic
pydantic

# LangGraph + Groq
langgraph
groq

# Embeddings
sentence-transformers

# Database
psycopg2-binary
pgvector

transformers
torch
sentencepiece
numpy

# LoRA fine-tuning
peft              # LoRA adapter library by HuggingFace
datasets          # HuggingFace datasets for training
accelerate        # training optimization
 
# MLOps
mlflow